+++ common/src/View/CellLayout.h
@@ -698,10 +698,10 @@ namespace TrenchBroom {
                 int newIndex = static_cast<int>(rowIndex) + offset;
                 if (newIndex < 0) {
                     while (newIndex < 0 && groupIndex > 0)
-                        newIndex += m_groups[--groupIndex].size();
+                        newIndex += static_cast<int>(m_groups[--groupIndex].size());
                 } else if (newIndex >= static_cast<int>(m_groups[groupIndex].size())) {
                     while (newIndex >= static_cast<int>(m_groups[groupIndex].size()) && groupIndex < m_groups.size() - 1)
-                        newIndex -= m_groups[groupIndex++].size();
+                        newIndex -= static_cast<int>(m_groups[groupIndex++].size());
                 }
 
                 if (groupIndex < m_groups.size()) {
