+++ common/src/View/TextureBrowserView.cpp
@@ -146,7 +146,7 @@ namespace TrenchBroom {
 
             if (m_group) {
                 for (const Assets::TextureCollection* collection : getCollections()) {
-                    layout.addGroup(collection->name(), fontSize + 2.0f);
+                    layout.addGroup(collection->name(), static_cast<float>(fontSize) + 2.0f);
                     for (Assets::Texture* texture : getTextures(collection))
                         addTextureToLayout(layout, texture, font);
                 }
